catalogwatch-ai/requirements.txt
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0
streamlit>=1.20
pytest>=7.0
